% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating.R
\name{rating}
\alias{rating}
\title{Rating
Basic hello world function to be called from the demo app}
\usage{
rating(mean = 20)
}
\arguments{
\item{mean}{Expected mean}
}
\description{
Rating
Basic hello world function to be called from the demo app
}
